refactor: extract imagesource children as common component
