+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/blocks/Image/ImageSource/ImageSource.tsx
@@ -1,12 +1,7 @@
-import Stack from '@mui/material/Stack'
-import Typography from '@mui/material/Typography'
 import { ReactElement } from 'react'
-import AddIcon from '@mui/icons-material/Add'
 import Card from '@mui/material/Card'
 import CardActionArea from '@mui/material/CardActionArea'
-import Box from '@mui/material/Box'
-import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline'
-import { ImageBlockThumbnail } from '../../../../../ImageBlockThumbnail'
+import { ImageSelection } from '../../../../../ImageLibrary/ImageSelection/ImageSelection'
 import { GetJourney_journey_blocks_ImageBlock as ImageBlock } from '../../../../../../../../__generated__/GetJourney'
 
 export interface ImageSourceProps {
@@ -37,36 +32,7 @@ export function ImageSource({
           display: 'flex'
         }}
       >
-        <Stack
-          direction="row"
-          sx={{ justifyContent: 'flex-start', alignItems: 'center' }}
-        >
-          <Box
-            sx={{
-              ml: 2,
-              mr: 4
-            }}
-          >
-            <ImageBlockThumbnail
-              selectedBlock={image !== undefined ? image : undefined}
-            />
-          </Box>
-          <Stack>
-            <Typography variant="subtitle2">
-              {image !== undefined ? image.id : 'Select Image'}
-            </Typography>
-            <Typography variant="caption">
-              {image !== undefined
-                ? `${image.width} x ${image.height} pixels`
-                : 'Upload your image'}
-            </Typography>
-          </Stack>
-        </Stack>
-        {image !== undefined ? (
-          <DeleteOutlineIcon color="primary" sx={{ mr: 2 }} />
-        ) : (
-          <AddIcon color="primary" sx={{ mr: 2 }} />
-        )}
+        <ImageSelection image={image} />
       </CardActionArea>
     </Card>
   )
