+++ apps/journeys-admin/src/components/Editor/ImageLibrary/ImageSelection/Index.ts
@@ -1,0 +1,1 @@
+export { ImageSelection } from './ImageSelection'
